#55 - Refactor authentication service: split auth logic into separate router and main server file
